+++ sdk/trafficmanager/azure-mgmt-trafficmanager/azure/mgmt/trafficmanager/_traffic_manager_management_client.py
@@ -7,14 +7,19 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING, cast
+from typing_extensions import Self
 
+from azure.core.pipeline import policies
 from azure.core.rest import HttpRequest, HttpResponse
+from azure.core.settings import settings
 from azure.mgmt.core import ARMPipelineClient
+from azure.mgmt.core.policies import ARMAutoResourceProviderRegistrationPolicy
+from azure.mgmt.core.tools import get_arm_endpoints
 
 from . import models as _models
 from ._configuration import TrafficManagerManagementClientConfiguration
-from ._serialization import Deserializer, Serializer
+from ._utils.serialization import Deserializer, Serializer
 from .operations import (
     EndpointsOperations,
     GeographicHierarchiesOperations,
@@ -24,31 +29,34 @@ from .operations import (
 )
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
+    from azure.core import AzureClouds
     from azure.core.credentials import TokenCredential
 
 
-class TrafficManagerManagementClient:  # pylint: disable=client-accepts-api-version-keyword
+class TrafficManagerManagementClient:
     """TrafficManagerManagementClient.
 
-    :ivar endpoints: EndpointsOperations operations
-    :vartype endpoints: azure.mgmt.trafficmanager.operations.EndpointsOperations
     :ivar profiles: ProfilesOperations operations
     :vartype profiles: azure.mgmt.trafficmanager.operations.ProfilesOperations
     :ivar geographic_hierarchies: GeographicHierarchiesOperations operations
     :vartype geographic_hierarchies:
      azure.mgmt.trafficmanager.operations.GeographicHierarchiesOperations
-    :ivar heat_map: HeatMapOperations operations
-    :vartype heat_map: azure.mgmt.trafficmanager.operations.HeatMapOperations
     :ivar traffic_manager_user_metrics_keys: TrafficManagerUserMetricsKeysOperations operations
     :vartype traffic_manager_user_metrics_keys:
      azure.mgmt.trafficmanager.operations.TrafficManagerUserMetricsKeysOperations
+    :ivar endpoints: EndpointsOperations operations
+    :vartype endpoints: azure.mgmt.trafficmanager.operations.EndpointsOperations
+    :ivar heat_map: HeatMapOperations operations
+    :vartype heat_map: azure.mgmt.trafficmanager.operations.HeatMapOperations
     :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The ID of the target subscription. Required.
     :type subscription_id: str
-    :param base_url: Service URL. Default value is "https://management.azure.com".
+    :param base_url: Service URL. Default value is None.
     :type base_url: str
+    :keyword cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
+     None.
+    :paramtype cloud_setting: ~azure.core.AzureClouds
     :keyword api_version: Api Version. Default value is "2022-04-01". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
@@ -58,29 +66,59 @@ class TrafficManagerManagementClient:  # pylint: disable=client-accepts-api-vers
         self,
         credential: "TokenCredential",
         subscription_id: str,
-        base_url: str = "https://management.azure.com",
+        base_url: Optional[str] = None,
+        *,
+        cloud_setting: Optional["AzureClouds"] = None,
         **kwargs: Any
     ) -> None:
+        _cloud = cloud_setting or settings.current.azure_cloud  # type: ignore
+        _endpoints = get_arm_endpoints(_cloud)
+        if not base_url:
+            base_url = _endpoints["resource_manager"]
+        credential_scopes = kwargs.pop("credential_scopes", _endpoints["credential_scopes"])
         self._config = TrafficManagerManagementClientConfiguration(
-            credential=credential, subscription_id=subscription_id, **kwargs
+            credential=credential,
+            subscription_id=subscription_id,
+            cloud_setting=cloud_setting,
+            credential_scopes=credential_scopes,
+            **kwargs
         )
-        self._client: ARMPipelineClient = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+
+        _policies = kwargs.pop("policies", None)
+        if _policies is None:
+            _policies = [
+                policies.RequestIdPolicy(**kwargs),
+                self._config.headers_policy,
+                self._config.user_agent_policy,
+                self._config.proxy_policy,
+                policies.ContentDecodePolicy(**kwargs),
+                ARMAutoResourceProviderRegistrationPolicy(),
+                self._config.redirect_policy,
+                self._config.retry_policy,
+                self._config.authentication_policy,
+                self._config.custom_hook_policy,
+                self._config.logging_policy,
+                policies.DistributedTracingPolicy(**kwargs),
+                policies.SensitiveHeaderCleanupPolicy(**kwargs) if self._config.redirect_policy else None,
+                self._config.http_logging_policy,
+            ]
+        self._client: ARMPipelineClient = ARMPipelineClient(base_url=cast(str, base_url), policies=_policies, **kwargs)
 
         client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
         self._serialize.client_side_validation = False
-        self.endpoints = EndpointsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.profiles = ProfilesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.geographic_hierarchies = GeographicHierarchiesOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.heat_map = HeatMapOperations(self._client, self._config, self._serialize, self._deserialize)
         self.traffic_manager_user_metrics_keys = TrafficManagerUserMetricsKeysOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
+        self.endpoints = EndpointsOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.heat_map = HeatMapOperations(self._client, self._config, self._serialize, self._deserialize)
 
-    def _send_request(self, request: HttpRequest, **kwargs: Any) -> HttpResponse:
+    def _send_request(self, request: HttpRequest, *, stream: bool = False, **kwargs: Any) -> HttpResponse:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
@@ -100,12 +138,12 @@ class TrafficManagerManagementClient:  # pylint: disable=client-accepts-api-vers
 
         request_copy = deepcopy(request)
         request_copy.url = self._client.format_url(request_copy.url)
-        return self._client.send_request(request_copy, **kwargs)
+        return self._client.send_request(request_copy, stream=stream, **kwargs)  # type: ignore
 
     def close(self) -> None:
         self._client.close()
 
-    def __enter__(self) -> "TrafficManagerManagementClient":
+    def __enter__(self) -> Self:
         self._client.__enter__()
         return self