+++ sdk/trafficmanager/azure-mgmt-trafficmanager/generated_samples/name_availability_v2_test_name_not_available_postexample23.py
@@ -1,3 +1,4 @@
+# pylint: disable=line-too-long,useless-suppression
 # coding=utf-8
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
@@ -7,6 +8,7 @@
 # --------------------------------------------------------------------------
 
 from azure.identity import DefaultAzureCredential
+
 from azure.mgmt.trafficmanager import TrafficManagerManagementClient
 
 """
@@ -35,6 +37,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/trafficmanager/resource-manager/Microsoft.Network/stable/2022-04-01/examples/NameAvailabilityV2Test_NameNotAvailable-POST-example-23.json
+# x-ms-original-file: specification/trafficmanager/resource-manager/Microsoft.Network/TrafficManager/stable/2022-04-01/examples/NameAvailabilityV2Test_NameNotAvailable-POST-example-23.json
 if __name__ == "__main__":
     main()